fix: stop worker when stop button is clicked
The bug was: after the button is clicked, SwingWorker.cancel(true) is
called, but doInBackground() still keeps running.
This fix adds checking of isCancelled() in doInBackground() to decide
whether it should continue.
Counting sequences is stoped immediately.
However, if MD5.calculateMD5() is started, it won't stop. We just let
it finish the calculation, but don't publish it's result. So user
won't see any change on the interface.